Move nfchelper
